agent/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
langgraph>=0.2.60
langchain-anthropic>=0.3.0
langchain-core>=0.3.40
langchain-mcp-adapters>=0.0.5
python-dotenv>=1.0.1
pydantic>=2.9.2
httpx>=0.27.2
